+++ packages/kit/src/views/Prime/hooks/usePrimeAuthV2.tsx
@@ -7,7 +7,7 @@ import backgroundApiProxy from '../../../background/instance/backgroundApiProxy'
 import { usePrivyUniversalV2 } from './usePrivyUniversalV2';
 
 export function usePrimeAuthV2() {
-  const [primePersistAtom] = usePrimePersistAtom();
+  const [user] = usePrimePersistAtom();
 
   const { logout, getAccessToken, isReady, authenticated } =
     usePrivyUniversalV2();
@@ -20,7 +20,9 @@ export function usePrimeAuthV2() {
   }, [logout]);
 
   return {
-    user: primePersistAtom,
+    isLoggedIn: user?.isLoggedIn,
+    isPrimeSubscriptionActive: user?.primeSubscription?.isActive,
+    user,
     logout: logoutWithApi,
     getAccessToken,
     isReady,
