+++ packages/kit/src/views/Prime/pages/PrimeDashboard/PrimeDashboard.tsx
@@ -1,10 +1,9 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 
 import { useIntl } from 'react-intl';
 import { StyleSheet } from 'react-native';
 
 import {
-  ActionList,
   Dialog,
   Icon,
   IconButton,
@@ -15,23 +14,17 @@ import {
   YStack,
   useSafeAreaInsets,
 } from '@onekeyhq/components';
-import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
-import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
-import { EWebEmbedRoutePath } from '@onekeyhq/shared/src/consts/webEmbedConsts';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import openUrlUtils from '@onekeyhq/shared/src/utils/openUrlUtils';
-import timerUtils from '@onekeyhq/shared/src/utils/timerUtils';
 
 import { PrimeLoginEmailDialogV2 } from '../../components/PrimeLoginEmailDialogV2';
+import { PrimePurchaseDialog } from '../../components/PrimePurchaseDialog/PrimePurchaseDialog';
 import { useFetchPrimeUserInfo } from '../../hooks/useFetchPrimeUserInfo';
 import { usePrimeAuthV2 } from '../../hooks/usePrimeAuthV2';
-import { usePrimePayment } from '../../hooks/usePrimePayment';
 
 import { PrimeBenefitsList } from './PrimeBenefitsList';
 import { PrimeDebugPanel } from './PrimeDebugPanel';
 import { PrimeLottieAnimation } from './PrimeLottieAnimation';
-import { PrimeSubscriptionPlans } from './PrimeSubscriptionPlans';
 import { PrimeUserInfo } from './PrimeUserInfo';
 
 function PrimeBanner() {
@@ -59,143 +52,48 @@ function PrimeBanner() {
 
 export default function PrimeDashboard() {
   const intl = useIntl();
-  const { user } = usePrimeAuthV2();
+  const { user, isLoggedIn, isPrimeSubscriptionActive } = usePrimeAuthV2();
   const { top } = useSafeAreaInsets();
-  const navigation = useAppNavigation();
   const { fetchPrimeUserInfo } = useFetchPrimeUserInfo();
   useEffect(() => {
     void fetchPrimeUserInfo();
   }, [fetchPrimeUserInfo]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [selectedPackageId, setSelectedPackageId] = useState<
-    string | undefined
-  >();
-  const {
-    presentPaywallNative,
-    purchasePaywallPackageWeb,
-    getPaywallPackagesWeb,
-  } = usePrimePayment();
-
-  const purchaseByWebview = useCallback(async () => {
-    navigation.popStack();
-    await timerUtils.wait(1000);
-    // purchase by webview
-    openUrlUtils.openUrlByWebviewPro({
-      url: '',
-      title: 'WebView',
-      isWebEmbed: true,
-      hashRoutePath: EWebEmbedRoutePath.primePurchase,
-      hashRouteQueryParams: {
-        primeUserId: user?.privyUserId || '',
-        primeUserEmail: user?.email || '',
-      },
-    });
-  }, [navigation, user?.privyUserId, user?.email]);
-
-  // TODO move to jotai context method
-  const doPurchase = useCallback(async () => {
-    try {
-      setIsLoading(true);
-      await timerUtils.wait(500);
-
-      if (!user?.isLoggedIn) {
-        const dialog = Dialog.show({
-          renderContent: (
-            <PrimeLoginEmailDialogV2
-              onComplete={() => {
-                void dialog.close();
-              }}
-            />
-          ),
-        });
-        return;
-      }
-
-      if (platformEnv.isNative) {
-        ActionList.show({
-          title: 'Purchase',
-          onClose: () => {},
-          sections: [
-            {
-              items: [
-                {
-                  label: 'Purchase by AppStore/GooglePlay',
-                  // description: 'Purchase by AppStore/GooglePlay',
-                  onPress: () => {
-                    void presentPaywallNative?.();
-                  },
-                },
-                {
-                  label: 'Purchase by Webview',
-                  // description: 'Purchase by Webview',
-                  onPress: () => {
-                    void purchaseByWebview();
-                  },
-                },
-              ],
-            },
-          ],
-        });
-        return;
-      }
-      if (selectedPackageId) {
-        await purchasePaywallPackageWeb?.({
-          packageId: selectedPackageId,
-          email: user?.email || '',
-        });
-        // await backgroundApiProxy.servicePrime.initRevenuecatPurchases({
-        //   privyUserId: user.privyUserId || '',
-        // });
-        // await backgroundApiProxy.servicePrime.purchasePaywallPackage({
-        //   packageId: selectedPackageId,
-        //   email: user?.email || '',
-        // });
-      }
-    } finally {
-      setIsLoading(false);
-      await fetchPrimeUserInfo();
-    }
-  }, [
-    user?.isLoggedIn,
-    user?.email,
-    selectedPackageId,
-    purchaseByWebview,
-    presentPaywallNative,
-    purchasePaywallPackageWeb,
-    fetchPrimeUserInfo,
-  ]);
 
   const shouldShowConfirmButton = useMemo(() => {
-    if (!user?.isLoggedIn) {
+    if (!isLoggedIn) {
       return true;
     }
-    if (user?.isLoggedIn && !user?.primeSubscription?.isActive) {
+    if (isLoggedIn && !isPrimeSubscriptionActive) {
       return true;
     }
     return false;
-  }, [user?.isLoggedIn, user?.primeSubscription]);
+  }, [isLoggedIn, isPrimeSubscriptionActive]);
 
-  const { result: paywallPackages } = usePromiseResult(async () => {
-    if (!platformEnv.isNative) {
-      return getPaywallPackagesWeb?.();
+  const subscribe = useCallback(() => {
+    if (!isLoggedIn) {
+      const loginDialog = Dialog.show({
+        renderContent: (
+          <PrimeLoginEmailDialogV2
+            onComplete={() => {
+              void loginDialog.close();
+            }}
+          />
+        ),
+      });
+
+      return;
     }
-  }, [getPaywallPackagesWeb]);
 
-  const subscriptionPlans = useMemo(() => {
-    if (
-      user?.isLoggedIn &&
-      // !user?.primeSubscription?.isActive &&
-      paywallPackages?.packages?.length
-    ) {
-      return (
-        <PrimeSubscriptionPlans
-          packages={paywallPackages?.packages}
-          onPackageSelected={setSelectedPackageId}
+    const purchaseDialog = Dialog.show({
+      renderContent: (
+        <PrimePurchaseDialog
+          onPurchase={() => {
+            void purchaseDialog.close();
+          }}
         />
-      );
-    }
-    return null;
-  }, [user?.isLoggedIn, paywallPackages]);
+      ),
+    });
+  }, [isLoggedIn]);
 
   return (
     <>
@@ -219,10 +117,7 @@ export default function PrimeDashboard() {
             >
               <PrimeLottieAnimation />
               <PrimeBanner />
-              {user?.isLoggedIn ? (
-                <PrimeUserInfo doPurchase={doPurchase} />
-              ) : null}
-              {subscriptionPlans}
+              {user?.isLoggedIn ? <PrimeUserInfo /> : null}
             </Stack>
 
             {platformEnv.isDev ? (
@@ -234,17 +129,10 @@ export default function PrimeDashboard() {
           </Page.Body>
 
           <Page.Footer
-            onConfirm={shouldShowConfirmButton ? doPurchase : undefined}
+            onConfirm={shouldShowConfirmButton ? subscribe : undefined}
             onConfirmText={intl.formatMessage({
               id: ETranslations.prime_subscribe,
             })}
-            confirmButtonProps={
-              shouldShowConfirmButton
-                ? {
-                    loading: isLoading,
-                  }
-                : undefined
-            }
           />
         </Page>
       </Theme>
